requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.3
easydict==1.13
fastapi==0.116.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.1
fsspec==2025.7.0
h11==0.16.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.5
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
onnx==1.19.0
onnxruntime==1.19.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
prettytable==3.16.0
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
simsimd==6.5.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
stringzilla==3.12.6
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.8.28
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.21
ultralytics-thop==2.0.16
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
wheel==0.45.1
